Initial emscripten port
The port is quite simple. A new platform is added in qcommon, then
we break the game loop for emscripten and finally get rid of all
GL extensions that make it crash.

This is all running with LEGACY_GL_EMULATION, though. A proper
OpenGL ES renderer would be needed for a real port.