Check for areas with invalid cluster 0 in AAS_AreaRouteToGoalArea()
Newer versions of BSPC such as 2.1h included with the Quake 3 GPL source
code create AAS files containing areas with cluster 0 if the area has no
reachabilities. Attempting to find the route to or from these areas
reportedly may cause the engine to crash due to an access violation.

The AAS files included with Quake 3 and Team Arena do not contain areas
with cluster 0.

I don't understand the AAS routing system well enough to know if areas
with cluster 0 is a bug in BSPC or some kind of optimization to ignore
unreachable areas. Either way the engine needs to handle the 15+ years
of add-on maps which have AAS areas with cluster 0.

Issue reported by Thomas Köppe.