Fix multiplayer server browser
Use dpmaster protocol so MFArena servers are listed on master server and
fetched correctly.

The reason using q3master protocol did not work before was the protocol
number. Setting PROTOCOL_LEGACY_VERSION to 68 would cause servers to be
listed and browser to work. Though then MFArena is not a separate game
from Quake 3.

To view Quake 3 servers in the browser (for UI testing) use:

    mfarena.x86 +set com_gamename Quake3Arena +set com_legacyprotocol 68