Add OpenGL ES 1.1 support to opengl1 renderer
This is loosely based on the OpenPandora port by ptitSeb. All images are
uploaded as RGBA because I have not added image format conversion
functions yet. There isn't software gamma correction so overbright may
not work (as is already the case for windowed mode on GNU/Linux).
Otherwise it should be fully working.

Some differences from the OpenPandora port:

* Doesn't use ifdefs for supporting OpenGL ES. The API is selected at
  run-time based on whether SDL2 defaults to OpenGL or OpenGL ES.
  A downside is that the original glBegin code may be easier to
  understand when porting the renderer to a different graphics API.
* Fixed debug drawing features - r_showtris, r_shownormals,
  and r_debugSurface.
* Enabling unsupported cvar settings (see README) automatically display
  a warning and disables the cvar.
* Reading pixels for screenshots and video recording uses RGBA/UBYTE or
  RGB/UBYTE if specified by GL_IMPLEMENTATION_COLOR_READ_FORMAT/TYPE.