OpenGL1: Use vertex arrays instead of glBegin
This is loosely based on the OpenPandora OpenGL ES port by ptitSeb.
Some of the changes were backported from the opengl2 renderer.

It runs faster, at least on modern hardware, and is requires for
OpenGL ES support.

r_primatives cvar values 1 and 3 still use glBegin as they have to use
it by definition.