Use poll() for network sockets on non-Windows
select() is limited to file descriptors 0 to 1023 which are shared by
sockets and files. The max file descriptors defaults to 1024 but the
plan is to increase it in the next commit to allow more pk3 files. It's
also possible to increase it using the command "ulimit -n 2048" before
running the game.

pk3 files are loaded before network is initialized and may cause the
network socket file descriptors to be over 1023 and thus invalid to
use with select().

(It's possible to increase the limit for select() on some platforms by
defining FD_SETSIZE but not Linux.)

Windows has separate sockets and file descriptors so it doesn't have
this issue. It doesn't have poll() so it continues to use select().